Fix collected bugs (...)
- crash when reference frame is not a sensor frame
- crash when deleting measurements without camera image
- numerical instabilities in board initialization
- type mismatch to message
- unused launch arg
- wrong marker scale
- typos

+++ extrinsic_sensor_calib/src/extrinsic_sensor_calib/types.h
@@ -80,7 +80,7 @@ using BoardID = std::string;
 using FrameID = std::string;
 using MeasurementIndex = std::size_t;
 
-using ReferencePointID = std::uint8_t;
+using ReferencePointID = std::int16_t;
 using CartesianCoordinate = Eigen::Vector3d;
 using LiDARBoardMeasurement = std::map<ReferencePointID, CartesianCoordinate>;
 using LiDARMeasurement = std::map<BoardID, LiDARBoardMeasurement>;
